backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.22
openai==1.10.0
numpy==1.26.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
httpx==0.26.0
yt-dlp
pydub
celery[redis]
redis
